Move to a two stage launching process for the CLI
Feedback from the OSGi specification committee pointed out that OSGi frameworks package their own API, and that the CLI should take all possible steps to avoid replacing that API in the classloader which loads the framework. This commit therefore restructures the build into units that have:

1. No OSGi dependencies - this includes basic implementations of Repositories that can be used without requiring the Feature Launcher API
2. No framework dependencies - this includes modules which need to decorate features, but not to interact with the framework
3. OSGi framework dependent modules - this includes modules which actually manage bundles and configurations

The launch process has been split into two stages. The first stage loads and decorates the feature. This also processes any extension which might select a framework. Once a framework is located it is loaded in its own classloader, and the second stage launcher proceeds as a child of that classloader. This ensures that the framework is not polluted with launcher dependencies, and that the second stage launcher can see the framework classes.

This change may seem very large, but the majority of changes are simple renames. The main code changes are in the CLI and launcher modules, which had to be change to the two-stage launch mechanism.

Signed-off-by: Tim Ward <timothyjward@apache.org>